+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2020-04-25
+
+- Remove `request.headers` from output
+- Add `os.environ` to output
+
+
 ## 2020-04-24
 
 - Project created